+++ changelog.md
@@ -35,6 +35,7 @@
 - `FileType` is now part of the prelude.
 - Implemented `Deref` and `DerefMut` for `Vector3D` properly, so accessing the methods of `nalgebra::Vector3` should be easier.
 - `AtomContainer::isin` (and consequently `System::group_isin`) should be now faster.
+- `Vector3D` now implements the `Display` trait.
 
 ***